+++ src/components/organisms/Settings/ShareSection/ShareSection.tsx
@@ -8,10 +8,10 @@ import UnshareIcon from '@/assets/icons/mail-off.svg?react';
 import ForwardIcon from '@/assets/icons/forward.svg?react';
 import useRepoSettingsStore from '@/stores/repoSettingsStore';
 import { useParams } from '@tanstack/react-router';
-import useShareRepositoryStatus from '@/hooks/useShareRepositoryStatus';
-import useRepositoryExit from '@/hooks/useRepositoryExit';
-import useGetRepositoryEntrycode from '@/hooks/useGetRepositoryEntrycode';
-import useRepositoryNewEntrycode from '@/hooks/useRepositoryNewEntrycode';
+import useShareRepositoryStatus from '@/hooks/common/useShareRepositoryStatus';
+import useRepositoryExit from '@/hooks/common/useRepositoryExit';
+import useGetRepositoryEntrycode from '@/hooks/common/useGetRepositoryEntrycode';
+import useRepositoryNewEntrycode from '@/hooks/common/useRepositoryNewEntrycode';
 import AlertDialogComponent from '@/components/molecules/AlertDialog/AlertDialog';
 import { useState } from 'react';
 import PasswordInput from '@/components/atoms/Input/PasswordInput';
